metaembedding/requirements.txt
pandas>=2.0.0
sentence-transformers>=5.0.0
umap-learn>=0.5.0
hdbscan>=0.8.0
matplotlib>=3.7.0
seaborn>=0.13.0

